Created 3 new service classes for handling subscriptions
This included splitting apart `EWSService`, placing anything to do with
elements in a new subclass: `EWSElementService`. This was done because
the subscription related classes didn't need to make use of elements, so
I felt it was worth simplifying the class. There will be some services
which don't need (and actually can't make use of) the elements.